+++ Koronavilkku/Koronavilkku/Environment.swift
@@ -5,6 +5,7 @@ import TrustKit
 struct Environment {
     let configuration: Configuration
     let batchRepository: BatchRepository
+    let efgsRepository: EFGSRepository
     let exposureRepository: ExposureRepository
     let municipalityRepository: MunicipalityRepository
 
@@ -23,7 +24,11 @@ extension Environment {
                                                   cache: LocalStore.shared,
                                                   storage: storage)
         
-        let exposureRepository = ExposureRepositoryImpl(exposureManager: ExposureManagerProvider.shared.manager,
+        let efgsRepository = EFGSRepositoryImpl(backend: backend,
+                                                storage: storage)
+
+        let exposureRepository = ExposureRepositoryImpl(efgsRepository: efgsRepository,
+                                                        exposureManager: ExposureManagerProvider.shared.manager,
                                                         backend: backend,
                                                         storage: storage)
         
@@ -33,6 +38,7 @@ extension Environment {
         
         return Environment(configuration: config,
                            batchRepository: batchRepository,
+                           efgsRepository: efgsRepository,
                            exposureRepository: exposureRepository,
                            municipalityRepository: municipalityRepository)
     }