+++ Koronavilkku/Koronavilkku/UI/Components/CardElement.swift
@@ -11,6 +11,15 @@ class CardElement: UIView {
         configureLayer()
     }
     
+    convenience init(embed view: UIView) {
+        self.init()
+        self.addSubview(view)
+        
+        view.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
+        }
+    }
+    
     override func layoutSubviews() {
         super.layoutSubviews()
         updateShadowPath()