+++ Koronavilkku/Koronavilkku/ViewController/ReportInfectionViewController.swift
@@ -8,11 +8,9 @@ class ReportInfectionViewController: UIViewController {
         case LockedText
         case DisabledText
         case ReportTitle
-        case ReportText
+        case ReportMessagePrimary
+        case ReportMessageSecondary
         case ReportButton
-        case ReportItemPrivacy
-        case ReportItemNotify
-        case ReportItemGuide
     }
     
     override func viewDidLoad() {
@@ -37,49 +35,16 @@ class ReportInfectionViewController: UIViewController {
         }
     }
     
-    private func pushToPublishTokensVC() {
-        self.pushToPublishTokensVC(with: nil)
-    }
-    
-    func pushToPublishTokensVC(with code: String?) {
-        guard LocalStore.shared.uiStatus != .apiDisabled else {
+    func startReportInfectionFlow(with code: String?) {
+        switch LocalStore.shared.uiStatus {
+        case .apiDisabled:
             Log.d("Cannot publish tokens because EN API is disabled")
-            return
-        }
-        
-        guard LocalStore.shared.uiStatus != .locked else {
+        case .locked:
             Log.d("User has already published tokens")
-            return
+        default:
+            self.navigationController?.present(ReportInfectionFlowViewController(publishToken: code),
+                                               animated: true)
         }
-
-        let publishTokensVC = PublishTokensViewController()
-        
-        if let code = code {
-            publishTokensVC.setCode(code)
-        }
-
-        let childNavController = UINavigationController(rootViewController: publishTokensVC)
-        childNavController.modalPresentationStyle = .fullScreen
-        
-        let appearance = UINavigationBarAppearance()
-        appearance.backgroundColor = UIColor.Secondary.blueBackdrop
-        appearance.shadowColor = .clear
-        appearance.titleTextAttributes = [
-            .font: UIFont.labelPrimary,
-            .foregroundColor: UIColor.Greyscale.black
-        ]
-        
-        let buttonAppearance = UIBarButtonItemAppearance()
-        buttonAppearance.normal.titleTextAttributes = [
-            .font: UIFont.labelPrimary,
-        ]
-
-        appearance.buttonAppearance = buttonAppearance
-        
-        childNavController.navigationBar.standardAppearance = appearance
-        childNavController.navigationBar.tintColor = UIColor.Primary.blue
-        
-        self.navigationController?.present(childNavController, animated: true)
     }
     
     private func showInstructions() {
@@ -88,34 +53,46 @@ class ReportInfectionViewController: UIViewController {
         navigationItem.title = Text.ReportTitle.localized
 
         let margin = UIEdgeInsets(top: 20, left: 20, bottom: 30, right: 20)
-        let contentView = view.addScrollableContentView(backgroundColor: UIColor.Secondary.blueBackdrop, margins: margin)
-        
+        let buttonMargin = UIEdgeInsets(top: 20, left: 20, bottom: 44, right: 20)
+        let contentView = view.addScrollableContentView(backgroundColor: UIColor.Secondary.blueBackdrop,
+                                                        margins: margin)
         var top = contentView.snp.top
+
+        let primaryMessage = UILabel(label: Text.ReportMessagePrimary.localized,
+                                     font: .heading4,
+                                     color: UIColor.Greyscale.black)
+
+        primaryMessage.numberOfLines = 0
+        top = contentView.appendView(primaryMessage, top: top)
         
-        let text = UILabel(label: Text.ReportText.localized, font: .bodyLarge, color: UIColor.Greyscale.black)
-        text.numberOfLines = 0
-        top = contentView.appendView(text, top: top)
+        let secondaryMessage = UILabel(label: Text.ReportMessageSecondary.localized,
+                                       font: .bodySmall,
+                                       color: UIColor.Greyscale.black)
         
-        let button = RoundedButton(title: Text.ReportButton.localized,
-                                   action: { [unowned self] in self.pushToPublishTokensVC() })
-        top = contentView.appendView(button, spacing: 30, top: top)
+        secondaryMessage.numberOfLines = 0
+        top = contentView.appendView(secondaryMessage, spacing: 20, top: top)
+
+        contentView.snp.makeConstraints { make in
+            // the content wrapper already contains the space between text and button top edge
+            make.bottom.equalTo(top).offset(buttonMargin.bottom + RoundedButton.height)
+        }
+
+        let fadeBlock = FadeBlock(color: UIColor.Secondary.blueBackdrop)
+        view.addSubview(fadeBlock)
+
+        fadeBlock.snp.makeConstraints { make in
+            make.bottom.equalTo(view.safeAreaLayoutGuide)
+            make.left.right.equalToSuperview()
+        }
+
+        let button = RoundedButton(title: Text.ReportButton.localized) { [unowned self] in
+            self.startReportInfectionFlow(with: nil)
+        }
+        
+        view.addSubview(button)
         
         button.snp.makeConstraints { make in
-            make.bottom.lessThanOrEqualTo(contentView.snp.bottom)
-        }
-        
-        var spacing: CGFloat = 30
-        var bottomAnchor = button.snp.bottom
-        
-        for listItem in [Text.ReportItemPrivacy, Text.ReportItemNotify, Text.ReportItemGuide] {
-            let bulletItem = BulletItem(text: listItem.localized)
-            top = contentView.appendView(bulletItem, spacing: spacing, top: top)
-            spacing = 10
-            bottomAnchor = bulletItem.snp.bottom
-        }
-        
-        contentView.snp.makeConstraints { make in
-            make.bottom.equalTo(bottomAnchor)
+            make.edges.equalTo(fadeBlock).inset(buttonMargin)
         }
     }