+++ Koronavilkku/Koronavilkku/ViewController/TestViewController.swift
@@ -17,16 +17,13 @@ class TestViewController: UIViewController {
     lazy var addExposureButton = self.createButton(title: "Add exposure", action: #selector(addExposure))
     lazy var addExposureDelayedButton = self.createButton(title: "Add exposure, delayed", action: #selector(addExposureDelayed))
     lazy var addLegacyExposureButton = self.createButton(title: "Add legacy exposure", action: #selector(addLegacyExposure))
-
     lazy var removeExposuresButton = self.createButton(title: "Remove exposures", action: #selector(removeExposures))
     lazy var radarStatus = self.createButton(title: "Radar status \(LocalStore.shared.uiStatus)", action: #selector(toggleRadarStatus))
-    
     lazy var resetOnboardingButton = self.createButton(title: "Reset onboarding", action: #selector(resetOnboarding))
     lazy var showExposureLogsButton = self.createButton(title: "Exposure logs", action: #selector(showExposureLogs))
-    
     lazy var updateMunicipalityListButton = self.createButton(title: "Update municipality list", action: #selector(updateMunicipalityList))
-    
     lazy var setLastExposureCheckButton = self.createButton(title: "Set exposure check date", action: #selector(setLastExposureCheck))
+    lazy var dumpEFGSCountriesButton = self.createButton(title: "Dump EFGS countries", action: #selector(dumpEFGSCountries))
     
     // force downcast, because we're using the internals here
     var batchRepository = Environment.default.batchRepository as! BatchRepositoryImpl
@@ -86,6 +83,7 @@ class TestViewController: UIViewController {
         appendButton(showExposureLogsButton)
         appendButton(updateMunicipalityListButton)
         appendButton(setLastExposureCheckButton)
+        appendButton(dumpEFGSCountriesButton)
         
         LocalStore.shared.$uiStatus.addObserver(using: { [weak self] in
             self?.radarStatus.setTitle("Radar status \(LocalStore.shared.uiStatus)", for: .normal)
@@ -213,6 +211,14 @@ class TestViewController: UIViewController {
     
     @objc func setLastExposureCheck() {
         LocalStore.shared.$dateLastPerformedExposureDetection.wrappedValue = Date().addingTimeInterval(30 - ExposureRepositoryImpl.manualCheckThreshold)
+    }
+    
+    @objc func dumpEFGSCountries() {
+        showDialog(
+            Environment.default.efgsRepository.getParticipatingCountries()?
+                .map { $0.localizedName }
+                .joined(separator: "\n") ?? "",
+            title: "EFGS Countries")
     }
     
     private func showDialog(_ message: String, title: String = "Error") {