+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-- New Checkbox and BulletItem components
+- New UI components: Checkbox, RadioButton, FadeBlock and BulletItem
+- EFGSRepository for managing the EFGS interop information: initially it just updates the country list on every app launch and provides a method for retrieving the participating countries.
 
 ### Changed
 - Improve accessibility by adding new header traits and hints, better checkbox handling
+- The current infection reporting functionality has been replaced with more detailed UI flow that optionally lets the user share their diagnosis keys with the EFGS. 
 
 ## 1.3.0